removed all warnings

+++ corewar/all_instructions/instructions_bis.c
@@ -10,10 +10,10 @@
 void exec_sub(vm_t *vm, champion_t *champion)
 {
     int byte_size = 5;
-    int a = champion->r[vm->arena[champion->pc + 2]];
-    int b = champion->r[vm->arena[champion->pc + 3]];
+    int a = champion->r[(int) vm->arena[champion->pc + 2]];
+    int b = champion->r[(int) vm->arena[champion->pc + 3]];
 
-    champion->r[vm->arena[champion->pc + 4]] = a - b;
+    champion->r[(int) vm->arena[champion->pc + 4]] = a - b;
     if (a - b == 0)
         champion->carry = 1;
     else
@@ -37,8 +37,8 @@ void exec_and(vm_t *vm, champion_t *champion)
     set_bit_int_at(&type, 0, get_bit_at(vm->arena[champion->pc + 1], 4));
     set_bit_int_at(&type, 1, get_bit_at(vm->arena[champion->pc + 1], 5));
     param2 = analyze_type(type, &act_pc, champion, vm);
-    champion->r[vm->arena[act_pc]] = param1 & param2;
-    if (param1 & param2 == 0)
+    champion->r[(int) vm->arena[act_pc]] = param1 & param2;
+    if ((param1 & param2) == 0)
         champion->carry = 1;
     else
         champion->carry = 0;
@@ -61,8 +61,8 @@ void exec_or(vm_t *vm, champion_t *champion)
     set_bit_int_at(&type, 0, get_bit_at(vm->arena[champion->pc + 1], 4));
     set_bit_int_at(&type, 1, get_bit_at(vm->arena[champion->pc + 1], 5));
     param2 = analyze_type(type, &act_pc, champion, vm);
-    champion->r[vm->arena[act_pc]] = param1 | param2;
-    if (param1 | param2 == 0)
+    champion->r[(int) vm->arena[act_pc]] = param1 | param2;
+    if ((param1 | param2) == 0)
         champion->carry = 1;
     else
         champion->carry = 0;
@@ -85,8 +85,8 @@ void exec_xor(vm_t *vm, champion_t *champion)
     set_bit_int_at(&type, 0, get_bit_at(vm->arena[champion->pc + 1], 4));
     set_bit_int_at(&type, 1, get_bit_at(vm->arena[champion->pc + 1], 5));
     param2 = analyze_type(type, &act_pc, champion, vm);
-    champion->r[vm->arena[act_pc]] = param1 ^ param2;
-    if (param1 ^ param2 == 0)
+    champion->r[(int) vm->arena[act_pc]] = param1 ^ param2;
+    if ((param1 ^ param2) == 0)
         champion->carry = 1;
     else
         champion->carry = 0;
